The Data Parallelism and ND-Range Kernel lessons duplicated a lot
of material, which became confusing when we tried presenting both
in a workshop. The duplication is unnecessary and also adds to the
maintenance cost.

Merge the "ND-Range Kernel" lesson into "Data Parallelism" by
moving all the unique slides from the latter into the former and
reordering so it has a reasonable flow. The exercises were also
similar, where "Data Parallelism" tested sycl::range + sycl::id
version and ND-Range tested sycl::range + sycl::item as well as
sycl::nd_range + sycl::nd_item. Merge the two by using most of
the code from the latter, but using sycl::id instead of sycl::item.
Merge the README instructions from both to retain full detail.

The top-level README is adjusted to renumber all lessons after the
removed one.

Constant memory is deprecated in SYCL 2020, so removed
it from the slides about the SYCL memory model.

+++ Code_Exercises/Data_Parallelism/source.cpp
@@ -44,11 +44,11 @@
 int main() {
   constexpr size_t dataSize = 1024;
 
-  float a[dataSize], b[dataSize], r[dataSize];
+  int a[dataSize], b[dataSize], r[dataSize];
   for (int i = 0; i < dataSize; ++i) {
-    a[i] = static_cast<float>(i);
-    b[i] = static_cast<float>(i);
-    r[i] = 0.0f;
+    a[i] = i;
+    b[i] = i;
+    r[i] = 0;
   }
 
   // Task: Compute r[i] = a[i] + b[i] in parallel on the SYCL device
@@ -56,5 +56,5 @@ int main() {
     r[i] = a[i] + b[i];
   }
 
-  SYCLACADEMY_ASSERT_EQUAL(r, [](size_t i) { return i * 2.0f; });
+  SYCLACADEMY_ASSERT_EQUAL(r, [](size_t i) { return i * 2; });
 }
